Changes to be committed: 	modified:   LICENSE 	new file:   mcdc.rb
This is the initial commit for mcdc.rb. The file contains the script to create test cases from a decision in such a way that the test cases fullfil the requiremets of MCDC.
This commit is used to make the tool available for the community.

The LICENSE file contains the correct company name and the period where in the script was created and tweaked.